Several aesthetic features added after algorithm completiona, notably - the ability to clear bool column of edge and node value data in a single click - as well as a warning in case this was performed on accident, a one stage and two stage option for the algorithm on Dr. Gilson's request, quick options for elucidating compound activation or inhibition, and a significantly faster paint nodes option for passthrough mapping  - takes 2s instead of 30s for large networks.
Last features left is to add style(update view) options to the algorithm (through a temporary class first, then to the real class). This will update the passthrough mapping as the algorithm proceeds. Also I still need to add the edge distinction button on the main menu, rather than the edge view. Finally, perhaps add a question mark for node analysis if the edge is zero - correlate this with passthrough mapping by mapping to an obscenely large number.

Signed-off-by: Aarya Venkat <Aarya Venkat>